requirements.txt
aiohttp==3.9.5
aiosignal==1.3.2
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.4.4
pillow==10.3.0
propcache==0.3.1
pyTelegramBotAPI==4.27.0
python-dotenv==1.0.1
python-telegram-bot==22.1
pytz==2025.2
requests==2.31.0
schedule==1.2.2
sniffio==1.3.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
yarl==1.20.0
tinkoff
tinkoff-investments==